Convert array type int to double

+++ 2016110002_HW6/src/Runner.java
@@ -5,8 +5,8 @@ public class Runner {
 	public static void main(String[] args) {
 		int n = 2;
 		int m = 3;
-		int[] objFuncCoeff = new int[n];
-		int[][] constCoeff = new int[m][n];
+		double[] objFuncCoeff = new double[n];
+		double[][] constCoeff = new double[m][n];
 		ArrayList<Double> objValue = new ArrayList<Double>();
 		
 		addNonNegativity(constCoeff);
@@ -14,7 +14,7 @@ public class Runner {
 		boolean isAllNeg = isObjFuncAllNeg(constCoeff[0]);
 		boolean unbounded = false;
 		// 0: constant, 1: x_1, ...
-		int[][] slackForm = getSlackForm(constCoeff);
+		double[][] slackForm = getSlackForm(constCoeff);
 		
 		while(!isAllNeg || unbounded) {
 			int maxCoeffIdx = getMaxCoeffIdx(slackForm[0]);
@@ -27,7 +27,7 @@ public class Runner {
 			
 			// (row, col) (const, coeff)
 			int tightConstIdx = getTightConstIdx(maxCoeffIdx, slackForm);
-			int[] rearranged = getRearranged(maxCoeffIdx, slackForm[tightConstIdx]);
+			double[] rearranged = getRearranged(maxCoeffIdx, slackForm[tightConstIdx]);
 			
 			replace(rearranged, slackForm);
 			
@@ -35,36 +35,36 @@ public class Runner {
 		}
 	}
 	
-	public static void addNonNegativity(int[][] constCoeff) {
+	public static void addNonNegativity(double[][] constCoeff) {
 	 
 	}
 	
-	public static boolean isObjFuncAllNeg(int[] objFunc) {
+	public static boolean isObjFuncAllNeg(double[] objFunc) {
 		// TODO: do not consider constant
 	}
 	
-	public static int[][] getSlackForm(int[][] constCoeff) {
+	public static double[][] getSlackForm(double[][] constCoeff) {
 		// slack form: [constraint + non-negativity][n variable + constant]
 		// [m+n][n+1]
 	}
 	
-	public static int getMaxCoeffIdx(int[] objFuncCoeff) {
+	public static int getMaxCoeffIdx(double[] objFuncCoeff) {
 		// TODO: do not consider constant
 	}
 	
-	public static boolean isUnbounded(int maxCoeffIdx, int[][] slackForm) {
+	public static boolean isUnbounded(int maxCoeffIdx, double[][] slackForm) {
 		
 	}
 	
-	public static int getTightConstIdx(int maxCoeffIdx, int[][] slackForm) {
+	public static int getTightConstIdx(int maxCoeffIdx, double[][] slackForm) {
 		
 	}
 	
-	public static int[] getRearranged(int maxCoeffIdx, int[] constraint) {
+	public static double[] getRearranged(int maxCoeffIdx, double[] constraint) {
 		// TODO: max coeff == 0
 	}
 	
-	public static void replace(int[] rearranged, int[][] slackForm) {
+	public static void replace(double[] rearranged, double[][] slackForm) {
 		
 	}
